[x86][acpi] Sketch AcpiEvaluateObject over FIDL support.
This CL does all the plumbing within the x86 board driver that's
necessary to be able to call AcpiEvaluateObject from a driver.
It also implements logic to check that a call is on an object that's the
child of a device.

Bug: 79172
Change-Id: I1648b3947add5b9984b3b77c7daa8192eee30bf1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551070
API-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>