[fint] Add support for RBE
This adds appropriate fields to context & set_artifacts proto and the
logic to wire them together. Note that while RBE and Goma behave
similarly, pre-deployed toolchain images are not currently deployed for
RBE and it is handling all uploads similar to Goma arbitrary toolchain
support so we do not need the same checks when using a custom toolchain
as Goma.

Bug: 80255
Change-Id: I774b09e1b93dbf91ca3b6d19aa44ab047458a713
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552128
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Nathan Mulcahey <nmulcahey@google.com>