[ffx] exit with non-zero status if experiment not opt-ed in
This changes ffx to exit with a non-zero status if someone tries to run
an experimental CLI, but did not opt-in to the experiment. This is
useful in scripting, since otherwise the script would assume the command
succeeded.

Fixed: 80069
Change-Id: I01fbbad0804e595d3696464c12a71565aff328d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551366
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Andy Weiss <dragonbear@google.com>