[fidlc] Fix dangerous identifiers in new syntax
Tests for `true` and `false` have been turned off, to be revisited by
fxbug.dev77561.  The resolves an edge case in the new synytax:

```
service Foo {
  bar client_end:true; // Should this be a literal 'true' or constant?
}
```

In the future we'll need to decide whether or not the `true` and `false`
constants may be overwritten (in which case they should always be parsed
as identifier constants), or whether such shadowing should be banned by
the compiled.

Additionally, other common identifiers are now tested as in the case
of constraints.

Bug: 77561
Test: fidldev test fidlc
Change-Id: I0c062c448b4ca47796fef601f79a181110a98f0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551816
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>