[fidl][rust] Validate that envelope byte count is a multiple of 8
This CL makes FIDL decoding Rust validate that the byte count field of
all envelopes is a multiple of 8. Previously this would trip an
assertion in the decoder, making it possible for a client to crash a
server simply by crafting a FIDL message in this way.

This also adds a GIDL test, so it's now tested for all bindings. It is
currently denylisted in Go, which returns the wrong error code.

Test: fx test fidl_rust_conformance_tests
Test: manually checked that new tests failed before fixes
Bug: 80162
Change-Id: I5c7085d10dd40d8bda34af17820c18721b09b028
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551921
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>