[debugger] Improve DWARF expression evaluation.
We have an async helper for evaluating DWARF expressions, the
AsyncDwarfExprEval. However, this is only used in one place. There is
another place where it has to go down to the underlying DwarfExprEval
and handle the tricky memory management issues.

We will be adding yet another similar place that requires bypassing the
async helper, so the design needs updating to accommodate all uses.

This separates the helper object into two layers, one just produces a
complete DwarfExprEval object with the result. The other automatically
converts that to a value.

There should be no behavior change.

Change-Id: I099562d65de2cb6cf81d9e3629baf015a78d42fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551379
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>