[lib/arch] Establish cache operation naming conventions
Whether the operation is for the local CPU or all of them is now
explicitly qualified with a qualifier of "Local" or "Global". (This
convention was previously present in part.)

Moreover, `cache_way_set_op()` is renamed to `data_cache_way_set_op()`
to clarify what cache it is operating on.

Change-Id: Id33b6a4b99daad4c5c3f92692248ea535f67cd3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552533
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>