[testrunner] Defer timeout enforcement to testers
Rather than enforcing it using a context deadline at the top level in
main.go. This is how testrunner was originally intended to enforce
timeouts, but it hasn't been working correctly since we added the
ability to repeatedly run a test up until a time limit, since we
enforced a context timeout with a duration equal to that enforced by the
tester.

The result is that tests will be canceled by the context deadline before
the tester hits its internal timeout, so we fail to do the proper
cleanup specific to each tester.

In commit 28ae1d4970a3b7806f70efe66a2cb555c3c7b6cb I assumed that code
was correct and refactored it to more clearly impose per-test timeouts
using a context deadline, without realizing that timeouts were already
being imposed by the tester structs. I don't think that made the problem
significantly worse, but it did make the code look more correct despite
the bug.

We still need to set a context timeout for serial tests to handle the
case where nothing gets printed to serial while the test is running, or
testrunner never sees the PASSED or FAILED result.

Previously the context timeout wrapped the entire Test() call and also
applied to the repeated attempts to start the serial command,
effectively limiting the number of attempts when -per-test-timeout is
set. Since the per-test timeout is now only applied to the
`runtests.TestPassed()` call that waits for a test result signature to
show up, we need an alternative way to limit the number of attempts we
make to run the test. So I just limited it to a constant 10 attempts.
This is much lower than the 60 attempts we would make previously on most
builders (5 minute total timeout, with a 5 second timeout for each
attempt) but I imagine 10 tries will probably be sufficient.

Also, set the context timeout to 30 seconds more than the value passed
to runtests' --timeout flag so that runtests actually has time to impose
the timeout before being canceled.

Bug: 80038
Change-Id: I1c01fd95c7ec586e5d996b38916f9f2166af6b0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551201
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>