[vmm] Pass ZBI_TYPE_MEM_CONFIG item to the x86 zircon guest
zbi_mem_range_t is now the preferred memory range metadata format and
early boot logic will soon take a hard dependency on it.

A ZbiMemRanges abstraction is introduced to encapsulate the
arch-specific manner in which ranges are specified, and in order to
share code with the E820 logic on the x86 side.

Bug: 65091
Bug: 80006
Change-Id: Ibe5cdee62bdb8518dd399b7382463eafc6242c68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552763
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>