[blobfs] Make metrics used by pager thread-safe.
Adds locks to the fields that are flushed to cobalt to make their access
thread-safe. Marking them atomic was a consideration since they are
largely just incremented, but the zx::ticks are not so easily converted
to atomic. The inspect fields do not require locks for modifying as
those fields are already thread-safe as they are guarded by the inspect
tree lock.

Additionally adding locks for the page-in frequencies, as they do some
operations on the map that would not be thread-safe.

Change-Id: Ic79285d0573bc6043dee53a1754920ea1b0bde04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552441
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>