[power] Implement ActivityHandler node
Implements a new node type called ActivityHandler. This node is
responsible for monitoring activity state changes from the
fuchsia.ui.activity.Provider service. When a change to the User Activity
state is observed, the ActivityHandler node sends a message to the
SystemProfileHandler node.

The new node is not instantiated on any device (will come in a later
CL).

Bug: 73863
Test: fx test power-manager-tests
Change-Id: If9edbabfbbe6f72dc851efebf2eecccf570cfd46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534236
Reviewed-by: Onath Claridge <claridge@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>