[fint] Add fvdl_sdk_test to never-affected tests
This is an e2e test that depends on the fvm so it's technically
affected (at least according to the build graph) by a large fraction of
fuchsia.git changes despite the test not actually being sensitive to
most such changes. So it's not useful to automatically multiply it on
those changes, and considering affected also makes it more likely that a
build will exceed the maximum of 3 affected tests allowed before no
affected tests will be multiplied in a build.

Change-Id: Ica773ca3a916884e0627d5dab87ba7427a01a730
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552982
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Yuan Zhi <yuanzhi@google.com>
Commit-Queue: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>