[mock_ddk] Convert ethertap driver tests to mock_ddk
This is part of a set of changes to discontinue use of the fake_ddk.
New drivers are instead recommended to use the mock_ddk or depend on
//src/devices/testing/no_ddk to get the necessary functions linked in.

This change also should address a flake with this test.

Bug: 80142

Testing: This is a change to test code.
Change-Id: I31802c333c4a5fc218cd467548c4a79a9d7ed675
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552682
Commit-Queue: Garratt Gallagher <garratt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Garratt Gallagher <garratt@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>