[ffx][repository] Register targets on start
This changes `Repo::register_target` to save which repositories we've
registered to config. Then in `Repo::start`, read the config and
re-register the repositories on the target. This allows ffx to make sure
that if the daemon is restarted that it makes sure the repository is still
set up.

In addition, this:

* Adds RepositoryManager::clear in order to drop all repositories when
  stopping the server.
* Changes FakeDaemon to support specifying the target nodename.
* Added a test to connect to the default target.

Change-Id: I0c74448210151a55175723b3c4fb909a9deeabe9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551308
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
API-Review: James Tucker <raggi@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>