[docs] Improve event capabilities docs
A lot of this will be rendered obsolete by the RFC
(https://fuchsia-review.googlesource.com/c/fuchsia/+/535692) currently
in review. But this represents the current state of the world and the
state we'll be for at least a few more months.

Fixed: 73003

Change-Id: I0232c2ce706c5bf66eac7e09d2ca78cbcaa58c3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547915
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>