[vmm][arm64] Always report that the UART TX FIFO is empty
PL011 gives separate TX FIFO "full" and "empty" bits; currently the
hypervisor always reports that both are false - so, in particular, that
the FIFO is never empty. It is a reasonable expectation for any driver
that this FIFO should eventually be empty - and we have a platform
driver that does expect this. Accordingly - and since incoming data in
the emulated UART is immediately processed - this change updates the
hypervisor to always report that the TX FIFO is empty.

Bug: 80376
Change-Id: I4dc26a09da60b1dbfcef104ca4e27a993de5f6e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553322
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>