[starnix] Restore signal mask properly
Previously the signal mask was restored when the task woke up inside of
sigsuspend. This would cause the signal dispatching logic to check the
signal against a different mask (the restored one) than was used to
determine whether or not to wake the task in the first place.

Change-Id: If93d5fdfbcfec3b82dc3ad7048b03018796eadcd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553146
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>