[fidl] Update fidlgen go library to use snake cased attribute names
This is done for consistency's sake with the rest of the code base.
Functionality is not affected, since all attribute names are converted
to lower-snake-case by the LookupAttribute method before comparison
anyway.  However, this is closer to our desired end state, so its nice
to get it done now.

Bug: 80353
Test: fidldev test fidlc
Change-Id: I6e289d1ae259ea4ea6b351f3fa6e7bd48208137c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553165
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>