[power] Implement InputSettingsHandler node
Implements a new node type called InputSettingsHandler. This node is
responsible for monitoring changes to input settings from the
fuchsia.settings.Input service. When a change to the input settings are
observed, the InputSettingsHandler node sends a message to the
SystemProfileHandler node.

The new node is not instantiated on any device (will come in a later
CL).

Bug: 73863
Test: fx test power-manager-tests
Change-Id: I188ef23c77c4199b0f9dde7b32b8ff8c76288a6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542343
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Patrick Shickel <pshickel@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>