[debugger] Add function step controls.
Generalize what to do when a new function is encountered during
stepping. This gives us a place to add logic to skip library functions
in the future.

Separates out the behavior for unsymbolized code from PLT trampolines.
It now always steps through the PLT code, while the unsymbolized
behavior applies to regular functions. The unsymbolized handling has
moved from a flag on "step" to a setting.

This replaces an old hack that stepped through the PLT trampolines by
detecting whether we were in unsymbolized code in a module that seems to
have symbols. We have since added real ELF symbols which allows us to
check for this case directly since the ELF PLT trampoline will have a
correspondingly annotated symbol.

Update the StepThreadController test to set up the symbols properly for
this test. Use the new InjectExceptionWithStack() to control symbols
more precisely.

This new behavior is incrementally better and is more flexible for
future additions. But it will still need to be significantly enhanced to
properly handle the general PLT problem (discussed in issue 80275).

Bug: 80258
Change-Id: I963f20c3bc0d13c3e79b7be27e4e1e9f9f272779
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552281
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>