[fdf][pbus] Replace coresident_device_index with primary_fragment
This was already updated in the libdriver API, this change just
propogates the change to the pbus API. One notable difference is that we
specify nullptr to specify that the primary_fragment is pdev.
Additionally, unless the primary fragment is set, we always spawn the
driver in a new driver host.

There should be not functional changes from this commit.

Change-Id: Ib04cf747ddd64ffe366bb31d8cf5e9e45e9d714a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551899
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>