[netemul] Return unescaped moniker
...from fuchsia.netemul/ManagedRealm.GetMoniker. Currently GetMoniker
returns a moniker with the ':' character escaped with a '\', which is
needed to use the moniker as a selector when reading inspect data.
Instead, return the actual moniker, and when using the moniker as a
selector, use a diagnostics utility to escape the necessary characters.

Test: fx test netemul-v2-tests
Test: fx test netstack-integration-tests-v2
Change-Id: Iccaadb3859f0ddefd82bcdbc769e8259dbceb6c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553307
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>