[wlan][policy] Fixes AP state reporting.
This change causes the AP state machine to retain a central
source of state information that can be used for providing
updates to listeners.  Listeners are now updated when the
state machine control channel is dropped and when the AP SME
event stream terminates.

Bug: 80292
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: Ib7aee598122817d9fc9063e58fe8a8d6eee66c4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552803
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>