[network/tests] Fix test_log_packets
The change in commit 6008bf32c3ca4b1fbf30a3148059020595d51ade introduced
a subtle error that did not cause the test to fail. Previously
`patterns` was used as a running list of log patterns that had *not* yet
been seen in log messages, and patterns would be removed once any log
message matched the pattern. The test passed once `patterns` was empty.

After this change, a given pattern is removed from `patterns` whenever a
log message is seen that doesn't match it. In other words, all it takes
for the test to pass is for a log message to be read that doesn't match
any of the patterns we're looking for.

This CL changes the test back to its intended behavior.

Test: fx test netstack-fidl-integration-test
Change-Id: Ie2fc8ddbb20169a89cd4eaac9750e737807ba84e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553541
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>