[gidl] Fix test: OptionalXUnionInStructPresentInvalidNumHandles
This test indicates there are 0 bytes and 1 handle, but a handle uses
4 bytes.

The previous test was problematic because 0 byte envelopes can't exist
in any present union in the v1 wire format. Once efficient envelopes
are in use, 0 byte envelopes are used to indicate presence, leading to
unexpected behavior when a 0 byte envelope appears for another reason.

Change-Id: Icdb3168efc190b37cb06745da668129be2b5c02c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553522
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>