[blobfs] Apply deadline profile to pager threads.
The switch to the new pager for blobfs lost the deadline profile on the
pager threads. This applies the deadline profile to all pager threads.

Test: manually verified that the function is called and succeeds.
Unfortunately, there is currently no way to query the profile on a
thread such that we can ensure this going forward.

Bug: 80357
Change-Id: I701a857b700ae8c92493b5b7acb0dcb7f5b1f7ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553209
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>