[debugger] Print return values before a stop.
Previously, the function return value from "finish" was printed before
the stop notification if it was synchronous (say, in a register), and
after if it was asynchronous (it was on the stack or printing it
required fetching memory). This is because the thread would continue its
work (either stopping or continuing) in parallel to the value being
evaluated.

This also prevented showing more return values that might happen in the
middle of stepping, say in "next".

This patch adds a system to evaluate these asynchronous expressions
before resuming the thread. It adds a new asynchronous phase before
resumption that avoids making all of the thread controllers
asynchronous. This allows the return value evaluation to have time to
finish before notifying the user or resuming the thread.

Bug: 5720
Change-Id: Ibf339a73dc9c8d925b06226b0ff7df967be31347
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553101
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>