[scrutiny] Add ffx scrutiny list packages.
This change adds a new command to the Scrutiny FFX interface which
allows you to list all of the fuchsia packages in the current build.
This is intented to be used as a discovery mechanism for the the
ffx scrutiny extract package command which allows you to extract a
package from a given package url.

Change-Id: I41cf37f05415e57c68649de2f92c7e0c87aab149
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551843
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Andy Weiss <dragonbear@google.com>