[helpdoc] add flag to escape jinja tags
Jinja tags in fx help docs cause a parser error in fuchsia.dev
publishing pipeline because the tags are attempted to be interpreted.
This CL adds a --escape-jinja flag that uses a mix of jinja escaping and
html escaping to avoid a parser error in includecode widget, which
expectes an argument and is validated for syntax issues before the verbatim
tags. Simple HTML escaping doesn't work because the reference docs in
fuchsia.dev is in a fenced code block, so the html entities would show
up as its code, not the expected symbol.

Change-Id: Ide6225605441a931f58ee3755b4651af32e2c7eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554197
Fuchsia-Auto-Submit: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Kelly-Ann Dolor <kadolor@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>