[starnix] Pass some of kill_test
This change passes all the signal tests that were blocked on bugs in
starnix's signal handling, as opposed to missing features (e.g., PID=1
workaround for missing /dev).

This fixes a bug where if a task was `sigsuspended` with a pending
signal, it would go to sleep and never wake. This would happen if the
`kill` came in during the `sigsuspend` execution.

It also adds support for waiting for a child process to exit (although
it does not encode the exit code correctly, yet).

Bug: 79289, 76976

Change-Id: Icd280e64ae7d3f8bf306a3d6400aaf4a34297dea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554192
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Barth <abarth@google.com>