[bt][mock-piconet-server] Add optional bt-rfcomm to topology
Provide API endpoint to optionally specify bt-rfcomm as an intermediary
component in the mock piconet. This matches production code in which the
RFCOMM component sits between bt-host and profiles.

Make a note that this should be used for integration tests of profiles
that require RFCOMM, and that integration tests for the bt-rfcomm
component itself should not use this hook.

Test: fx test mock-piconet-client-v2-tests
Bug: 79742
Change-Id: Iaa8b890425c02c7535b3124f7b0876986a62c7ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553045
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>