[kernel][boot-options]: kernel.serial.
There is some 'shared' logic between physboot and kernel proper now.
The logic is 'whether i found an entry for kernel.serial' in the command
line argument.
This is captured as 'gBootOption->serial_source' which is available
until full unification is achieved.

The kernel then will use this information to either parse something from
the commandline or the zbi, with its own logic.

This is necessary because Boot Options, at this time, does not provide
any information on whether the value is default, zbi, or command line.
So the kernel could not make the decision whether to fall back to the
zbi or not. Now it doesnt fallback it just uses whatever physboot
used(nothing, commandline or zbi).

Bug: 53594
Test: CQ
Change-Id: I9e0d14fad1798f2a4acde09abb678d2f28dacff0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546582
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>