[flatland] make OnError separate
As part of the transition from OnPresentProcessed() to
OnNextFrameBegin() semantics, we should make the Error portion of the
event separate, as errors are on a per-present basis, and
OnNextFrameBegin will not necessarily fire 1:1 with previous presents.
It also helps simplify the logic conceptually for clients.

Test: changed unittests
Bug: 79131
Change-Id: Ifd66d4463cda90cc57a3e1f7465a41e873294392
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553310
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Mikael Pessa <mikaelpessa@google.com>
API-Review: John Bauman <jbauman@google.com>
Commit-Queue: Felipe Archondo <felipearchondo@google.com>