[bt][avdtp] Have AVDTP commands send immediately
Update the API to send the command before returning the result, a Future
which you can poll to receive the response.

This makes it possible to send a command that you don't care to get the
result from (i.e. an Abort or a Suspend that should only result in an
acknowledgement) without needing to poll the Future.

Previously commands would also borrow the peer object and arguments
until the response was received. Because they are only used for
constructing the command to be sent, this also removes that borrow;
response Futures have unrelated lifetimes.

Test: added command_response_can_be_discarded; fx test bt-avdtp-tests
Bug: 77017

Change-Id: If46b28cc80b0700b84c15cdcc7234804e5ad6133
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553970
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>