[network/tests] Only read logs from the netstack under test
Currently test_log_packets uses fuchsia.logger/LogListenerSafe to listen
to syslogs and verify the netstack appropriately logs when it sends/
receives packets according to its configuration. However, the
fuchsia.logger protocols don't include moniker information, and since
this test is connecting to the system-wide logs, it's reading logs from
every netstack that happens to be under test while the test is running.

This CL changes the test to use fuchsia.diagnostics/ArchiveAccessor,
which includes moniker information in logs. Then we filter the log
messages to exclude those not originating from the netstack spawned for
this test, so we correctly only pass/fail based on those logs.

Additionally, tests use the debug netstack, which is already configured
to log packets, so this test was not actually testing the
`fuchsia.net.stack/Log.SetLogPackets` knob. To fix this, this CL teaches
netemul how to override program args for components, and removes the
`--log-packets` arg from the netstack for this test.

Fixed: 80036
Test: fx test netstack-fidl-integration-test
Change-Id: I48a76c8cbe6c435e24a70ff9dbf079a7ce04e4c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553966
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>