[docs][bt] Update persistent data schema doc
A variety of small cleanups so that the document accurately describes
the current schema used. Aside from small naming changes, elaborations,
or fixes to the examples, the most significant difference is the use of
a "PeerKey" type to represent a key with security properties, instead
of security-property-having and security-property-less Keys sharing the
same type, which affected a number of types which had Key components.

Fixed: 79702
Change-Id: Ic440a15c09e102aec1d2ec99c0252118189576de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553666
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>