[build] Make clang_prefix not rebased by default.
This CL changes the definition of the 'clang_prefix' variable
to a source-absolute or system-absolute value (in the case where
it is redefined through args.gn). The fact that it was previously
rebased was a source of confusion.

It also adds rebased_clang_prefix which is the same path rebased
to root_build_dir, as it is used in multiple places.

For testing, the content of all .ninja files before/after this
CL were compared. The only difference is the path of the clang-format
tool used by the fidling_toolchain.ninja file which is not relative
to the root build dir, instead of being absolute.

Note: //third_party/go/BUILD.gn still uses rebase_path(clang_prefix)
which will expand to the same absolute path before/after this CL.
Fixing this to use relative paths instead requires a different CL in
the //third_party/go.git repository.

Bug: None

Change-Id: If34b508ce2781cd9a806b0b9cf637ae46a00c8f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554562
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>