[audio_core] Rename AudioClockManager -> AudioClockFactory
This is a nit:

Making this change because in VNext, I'd like to standardize terminology
and I've started using "Factory" for "thing that makes other things".
"Factory" is the standard term for this. A "Manager" typically manages
the lifecycle of objects which may be created elsewhere.

Change-Id: I40895b74ac1574729c669886753ddf10f87be4d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543101
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>