[build] Make rustc_prefix and rust_sysroot source-absolute
This CL makes the value of the 'rustc_prefix' and 'rust_sysroot'
global path source-absolute or system-absolute (if redefined through
a build argument).

This makes it more consistent with 'clang_prefix', and less surprising
overall.

This also adds 'rebased_rustc_prefix', which is the same path rebased
to 'root_build_dir', useful in string expansions that are passed as
command line arguments.

For testing, the content of out/default generated by 'fx gen' was
compared before/after this CL was applied. It is identical, except
for the fact that --toolchain-lib-dir PATH now uses a shorter path
that replaces ROOT/bin/../lib with ROOT/lib.

Bug: None

Change-Id: Ie5ff5ad6bd950dfc3122522ff972b9f496263306
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554563
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: David Fang <fangism@google.com>