[ffx] ffx e2e experiment tests
This creates an experimental plugin that can be called from an e2e test
to ensure that the code generation around experiments is working as
intended. We test that you get a non-successful exit code and a message
when you try to run a plugin that is a non-enabled experiment. And we
test that enabling that plugin allows you to run it successfully.

Fixes https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=80133

Change-Id: Ia8b8eea18d615dae5defc1c874dec10f643b64c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554826
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>