[docs][components] Improve `eager` documentation
`eager` has some sharp edges around package sets. Most of these
types of components should be in the base package set. At a
minimum the `eager` component should have the same availability
as its parent, which is to say it should be in the same package
set.

Change-Id: Ibcb95ab443d95c687b8faad84193b08bcf88c115
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551446
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Dave Smith <smithdave@google.com>
Reviewed-by: Kyo Lee <kyol@google.com>