[kernel][boot-options]: Reland "kernel.serial."
This is a reland of c7e7bfb737c5c5d33db2e65a35ab1403054e57cb

There was the wrong assumption that ParseBootOptions was called only once,
but it is called once per zbi item cmdline. This means that the serial_source
needs to be preserved across calls.

This is done by using OptionSource::kDefault as the only state that can transition
to OptionSource::kZbi.

The original bug is as follows:
zbi cmdline item 1 : "serial=.... k1=v2"
zbi cmdline item 2 : "....."

When item 1 is processed, the serial source becomes kCmdLine, but when the second
one is parsed, it is resetted to kZbi, and because there is no serial option in the
second item, then the source is incorrectly set to kZbi.

Original change's description:
> [kernel][boot-options]: kernel.serial.
>
> There is some 'shared' logic between physboot and kernel proper now.
> The logic is 'whether i found an entry for kernel.serial' in the command
> line argument.
> This is captured as 'gBootOption->serial_source' which is available
> until full unification is achieved.
>
> The kernel then will use this information to either parse something from
> the commandline or the zbi, with its own logic.
>
> This is necessary because Boot Options, at this time, does not provide
> any information on whether the value is default, zbi, or command line.
> So the kernel could not make the decision whether to fall back to the
> zbi or not. Now it doesnt fallback it just uses whatever physboot
> used(nothing, commandline or zbi).
>
> Bug: 53594
> Test: CQ
> Change-Id: I9e0d14fad1798f2a4acde09abb678d2f28dacff0
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546582
> Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
> Reviewed-by: Nick Maniscalco <maniscalco@google.com>

Bug: 53594
Change-Id: Ib8b30fc135775df92203e1ba15c7f4c75adeb382
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554821
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>