[kernel][vm] Fix off by one error in enumeration
VmAddressRegion::EnumerateChildrenInternalLocked has an off by one in
its enumeration where it could eroneously call the on_mapping callback
for a VmMapping that was not in the range.

This is not normally triggered since the start of the iteration uses
subregions_.IncludedOrHigher to begin iteration at a valid mapping. But
if walking down into child vmars it is now possible to have mappings
that occur prior to the range be considered. The included test
constructs such a hierarchy as demonstration.

Bug: 80034
Change-Id: I6921de954f757a1a10284db3a63c68d9b65193c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555763
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>